+++ cmd/influxd/run/server.go
@@ -2,7 +2,9 @@ package run
 
 import (
 	"bytes"
+	"encoding/json"
 	"fmt"
+	"io/ioutil"
 	"log"
 	"net"
 	"net/http"
@@ -72,8 +74,10 @@ type Server struct {
 
 	Monitor *monitor.Monitor
 
-	// Server reporting
+	// Server reporting and registration
 	reportingDisabled bool
+	enterpriseURL     string
+	enterpriseToken   string
 
 	// Profiling
 	CPUProfile string
@@ -100,6 +104,8 @@ func NewServer(c *Config, buildInfo *BuildInfo) (*Server, error) {
 		Monitor: monitor.New(c.Monitor),
 
 		reportingDisabled: c.ReportingDisabled,
+		enterpriseURL:     c.EnterpriseURL,
+		enterpriseToken:   c.EnterpriseToken,
 	}
 
 	// Copy TSDB configuration.
@@ -397,6 +403,11 @@ func (s *Server) Open() error {
 			go s.startServerReporting()
 		}
 
+		// Register server
+		if err := s.registerServer(); err != nil {
+			log.Printf("failed to register server: %s", err.Error())
+		}
+
 		return nil
 
 	}(); err != nil {
@@ -506,6 +517,59 @@ func (s *Server) reportServer() {
 
 	client := http.Client{Timeout: time.Duration(5 * time.Second)}
 	go client.Post("http://m.influxdb.com:8086/db/reporting/series?u=reporter&p=influxdb", "application/json", data)
+}
+
+// registerServer registers the server on start-up.
+func (s *Server) registerServer() error {
+	if s.enterpriseToken == "" {
+		return nil
+	}
+
+	clusterID, err := s.MetaStore.ClusterID()
+	if err != nil {
+		log.Printf("failed to retrieve cluster ID for registration: %s", err.Error())
+		return err
+	}
+
+	hostname, err := os.Hostname()
+	if err != nil {
+		return err
+	}
+
+	j := map[string]interface{}{
+		"cluster_id": fmt.Sprintf("%d", clusterID),
+		"server_id":  fmt.Sprintf("%d", s.MetaStore.NodeID()),
+		"host":       hostname,
+		"product":    "influxdb",
+		"version":    s.buildInfo.Version,
+	}
+	b, err := json.Marshal(j)
+	if err != nil {
+		return err
+	}
+
+	url := fmt.Sprintf("%s/api/v1/servers?token=%s", s.enterpriseURL, s.enterpriseToken)
+	go func() {
+		client := http.Client{Timeout: time.Duration(5 * time.Second)}
+		resp, err := client.Post(url, "application/json", bytes.NewBuffer(b))
+		if err != nil {
+			log.Printf("failed to register server with %s: %s", s.enterpriseURL, err.Error())
+			return
+		}
+		defer resp.Body.Close()
+
+		if resp.StatusCode == http.StatusCreated {
+			return
+		}
+
+		body, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			log.Printf("failed to read response from registration server: %s", err.Error())
+			return
+		}
+		log.Printf("failed to register server with %s: received code %s, body: %s", s.enterpriseURL, resp.Status, string(body))
+	}()
+	return nil
 }
 
 // monitorErrorChan reads an error channel and resends it through the server.